+++ matlab/fft_filter.m
@@ -12,8 +12,18 @@ function mrWav_filt = fft_filter(mrWav, P, vcMode)
 
 if nargin<3, vcMode = ''; end
 if isempty(vcMode), vcMode = 'bandpass'; end
+fDebug = 0;
+
+if P.fGpu && fDebug
+    mrWav = gather(mrWav);
+    P.fGpu = 0;
+    fGpu_out = 1;
+else
+    fGpu_out = 0;
+end
 
 NSKIP_MAX = 2^19; % fft work length
+scale_filter = get_set_(P, 'scale_filter', 1);
 nPad = 300;
 [nT, nC] = size(mrWav);
 nSkip = min(nT, NSKIP_MAX);
@@ -22,22 +32,23 @@ nSkip = min(nT, NSKIP_MAX);
 if isempty(freqLim), mrWav_filt = mrWav; return; end
 
 try
-    mrWav_filt = zeros(size(mrWav), 'like', mrWav); catch
-    mrWav_filt = zeros(size(mrWav), class_(mrWav)); mrWav_filt
+    mrWav_filt = zeros(size(mrWav), 'like', mrWav); 
+catch
+    mrWav_filt = zeros(size(mrWav), class_(mrWav)); 
     fGpu = 0;
 end
 switch lower(vcMode)
-    case 'bandpass', fh_filter = @fft_bandpass_;
-    case 'highpass', fh_filter = @fft_highpass_;
-    case 'lowpass', fh_filter = @fft_lowpass_;
-    case {'fftdiff', 'banddiff'}, fh_filter = @fft_diff_;
+    case 'bandpass', fh_make_filter = @fft_bandpass_;
+    case 'highpass', fh_make_filter = @fft_highpass_;
+    case 'lowpass', fh_make_filter = @fft_lowpass_;
+    case {'fftdiff', 'banddiff'}, fh_make_filter = @fft_diff_;
     case 'wiener'
-        fft_wiener_(mrWav, P); % build cache if not built yet
-        fh_filter = @fft_wiener_;
+        vrFilt_ = fft_wiener_(mrWav, P); % build cache if not built yet
+        fh_make_filter = @fft_wiener_;
     otherwise, error(['fftfilt_: invalid option: ', vcMode]);
 end
 if ~fGpu, mrWav = gather_(mrWav); end
-fh_filt = @(x,f)real(ifft(bsxfun(@times, fft(single(x)), f)));
+fh_filter = @(x,f)real(ifft(bsxfun(@times, fft(single(x)), f)));
 n_prev = nan;
 fprintf('Running fft filter (%s)\n\t', vcMode); t1=tic;
 for iStart = 1:nSkip:nT
@@ -56,14 +67,19 @@ for iStart = 1:nSkip:nT
     [mrWav1, fGpu] = gpuArray_(mrWav(vi1,:), fGpu);
     n1 = size(mrWav1,1);
     if n1 ~= n_prev
-        vrFilt1 = fh_filter(n1, freqLim, freqLim_width, sRateHz);
+        vrFilt1 = fh_make_filter(n1, freqLim, freqLim_width, sRateHz);
         vrFilt1 = gpuArray_(vrFilt1, fGpu);
+        if scale_filter ~= 1
+            vrFilt1 = vrFilt1 * scale_filter;
+        end
         n_prev = n1;
     end    
-    try
-        mrWav1 = fh_filt(mrWav1, vrFilt1);  
-    catch
-        mrWav1 = fh_filt(gather_(mrWav1), vrFilt1);  
+    if ~isempty(vrFilt1)        
+        try
+            mrWav1 = fh_filter(mrWav1, vrFilt1);  
+        catch
+            mrWav1 = fh_filter(gather_(mrWav1), vrFilt1);  
+        end
     end
     if ~isGpu_(mrWav_filt)
         mrWav_filt(iStart:iEnd,:) = gather_(mrWav1(nPad+1:end-nPad,:));
@@ -71,9 +87,12 @@ for iStart = 1:nSkip:nT
         mrWav_filt(iStart:iEnd,:) = mrWav1(nPad+1:end-nPad,:);
     end
     mrWav1 = []; % clear memory
-%     fprintf('.');
+    fprintf('.');
 end
 if ~isGpu_(mrWav), mrWav_filt = gather_(mrWav_filt); end
+if fGpu_out
+    mrWav_filt = gpuArray(mrWav_filt);
+end
 fprintf('\n\ttook %0.1fs (fGpu=%d)\n', toc(t1), fGpu);
 end %func
 
@@ -87,6 +106,7 @@ function filt = fft_bandpass_(N, freqLim, freqLim_width, sRateHz)
 % sRateHz: sampling rate
 % freqLim: frequency limit, [f_lo, f_hi]
 % freqLim_width: frequency transition width, [f_width_lo, f_width_hi]
+if isempty(freqLim), freqLim = [nan, nan]; end
 [flo, fhi] = deal(freqLim(1), freqLim(2));
 [fwid_lo, fwid_hi] = deal(freqLim_width(1), freqLim_width(2));
 
@@ -160,14 +180,16 @@ function vrFilter_wiener = fft_wiener_(N, freqLim, freqLim_width, sRateHz)
 % freqLim_width: frequency transition width, [f_width_lo, f_width_hi]
 
 persistent P_ filt0_
-if nargin==2, [mr_, P_] = deal(N, freqLim); end
 
 % computes the cache only if the cache is abscent
-if isempty(filt0_)
-    mr1 = single(mr_(1:end/8, :)); % make a smaller copy
-    if 1
-        mr1 = fft_filter(mr1, P_, 'highpass');
+if nargin==2
+    if ~isempty(filt0_)
+        vrFilter_wiener = filt0_;
+        return;
     end
+    [mr_, P_] = deal(N, freqLim);
+    mr1 = single(mr_(1:end/8, :)); % make a smaller copy
+    mr1 = fft_filter(mr1, P_, 'bandpass');
     % reshape to the spike size and perform fft
     nSamples = (diff(P_.spkLim_raw) + 1)*2; %*2;
     mr2 = reshape_(mr1(:), nSamples);
@@ -184,22 +206,27 @@ if isempty(filt0_)
         case 1, fh1 = @(x)median(abs(mrFft(:,x)),2);
     end
     vrFft_noise = fh1(vr2<=threshLim(1));
-    vrFft_signal = fh1(vr2>=threshLim(2));
-    switch 1
-        case 2, scale = quantile(vrFft_signal ./ vrFft_noise, 1/8);
-        case 1, scale = min(vrFft_signal(2:end)) / min(vrFft_noise(2:end));
-    end
-
-    vrFft_signal = abs(vrFft_signal(:) / scale - vrFft_noise(:));
-    switch 1
-        case 2, vrFilt = vrFft_signal ./ (vrFft_signal + vrFft_noise);
-        case 1, vrFilt = vrFft_signal.^2 ./ (vrFft_signal.^2 + vrFft_noise.^2);
-    end
+    vrFft_noise_signal = fh1(vr2>=threshLim(2));
+    vrFilt = fh_rect(vrFft_noise_signal.^2 - vrFft_noise.^2) ./ vrFft_noise_signal.^2;
+    vrFilt(vrFilt<0) = 0;
+%     switch 3
+%         case 3, scale = 1;
+%         case 2, scale = quantile(vrFft_signal ./ vrFft_noise, 1/4);
+%         case 1, scale = min(vrFft_signal(2:end)) / min(vrFft_noise(2:end));
+%     end
+% 
+%     vrFft_signal = vrFft_signal(:) / scale - vrFft_noise(:);
+%     vrFft_signal(vrFft_signal<0) = 0;
+%     switch 1
+%         case 2, vrFilt = vrFft_signal ./ (vrFft_signal + vrFft_noise);
+%         case 1, vrFilt = vrFft_signal.^2 ./ (vrFft_signal.^2 + vrFft_noise.^2);
+%     end
     vrFilt(1) = vrFilt(end);
     filt0_ = vrFilt;
     fprintf(2, 'fft_filter: wiener filter recomputed `filt0_`\n');
+    vrFilter_wiener = vrFilt;
+    return;
 end
-if nargout==0, return; end
 
 
 nSamples = numel(filt0_);
@@ -213,8 +240,8 @@ switch 2
         vrFilter = vcFilter_diff .* vrFilter_wiener;
         vrFilter = vrFilter / max(vrFilter) * 2;
     case 2 %combine with High-pass
-        vrFilter_hp = fft_highpass_(N, P_.freqLim, P_.freqLim_width, P_.sRateHz);
-        vrFilter = vrFilter_hp(:) .* vrFilter_wiener;
+        vrFilter_bp = fft_bandpass_(N, P_.freqLim, P_.freqLim_width, P_.sRateHz);
+        vrFilter = vrFilter_bp(:) .* vrFilter_wiener;
     case 1
         vrFilter = vrFilter_wiener;
 end
@@ -365,4 +392,58 @@ vrFilt12 = (vrFilt(vi1) + vrFilt(vi2))/2;
 vrFilt1(vi1) = vrFilt12;
 vrFilt1(vi2) = vrFilt12;
 % vrFilt1(1) = 0; % zero DC gain
+end %func
+
+
+%--------------------------------------------------------------------------
+% 17/9/13 JJJ: Behavior changed, if S==[], S0 is loaded
+function val = get_set_(S, vcName, def_val)
+% set a value if field does not exist (empty)
+
+if isempty(S), S = get(0, 'UserData'); end
+if isempty(S), val = def_val; return; end
+if ~isstruct(S)
+    val = []; 
+    fprintf(2, 'get_set_: %s must be a struct\n', inputname(1));
+    return;
+end
+val = get_(S, vcName);
+if isempty(val), val = def_val; end
+end %func
+
+
+%--------------------------------------------------------------------------
+% 1/31/2019 JJJ: get the field(s) of a struct or index of an array or cell
+function varargout = get_(varargin)
+% same as struct_get_ function
+% retrieve a field. if not exist then return empty
+% [val1, val2] = get_(S, field1, field2, ...)
+% [val] = get_(cell, index)
+
+if nargin==0, varargout{1} = []; return; end
+S = varargin{1};
+if isempty(S), varargout{1} = []; return; end
+
+if isstruct(S)
+    for i=2:nargin
+        vcField = varargin{i};
+        try
+            varargout{i-1} = S.(vcField);
+        catch
+            varargout{i-1} = [];
+        end
+    end
+elseif iscell(S)
+    try    
+        varargout{1} = S{varargin{2:end}};
+    catch
+        varargout{1} = [];
+    end
+else
+    try    
+        varargout{1} = S(varargin{2:end});
+    catch
+        varargout{1} = [];
+    end
+end
 end %func